+++ projects/start-os/web/ui/src/app/routes/portal/components/interfaces/addresses/plugin/actions.component.ts
@@ -236,7 +236,7 @@ export class PluginActionsComponent {
       actionInfo: { id: actionId, metadata },
       prefill: {
         urlPluginMetadata: {
-          packageId: this.packageId() || null,
+          packageId: this.packageId(),
           hostId: addressInfo.hostId,
           interfaceId: iface.id,
           internalPort: addressInfo.internalPort,