+++ shared-libs/crates/start-core/src/service/effects/net/bind.rs
@@ -1,4 +1,5 @@
 use crate::net::host::binding::{BindId, BindOptions, NetInfo};
+use crate::net::host::host_for;
 use crate::service::effects::prelude::*;
 use crate::{HostId, PackageId};
 
@@ -123,19 +124,8 @@ pub async fn get_service_port_forward(
 
     let package_id = package_id.unwrap_or_else(|| context.seed.id.clone());
 
-    Ok(context
-        .seed
-        .ctx
-        .db
-        .peek()
-        .await
-        .as_public()
-        .as_package_data()
-        .as_idx(&package_id)
-        .or_not_found(&package_id)?
-        .as_hosts()
-        .as_idx(&host_id)
-        .or_not_found(&host_id)?
+    let mut db = context.seed.ctx.db.peek().await;
+    Ok(host_for(&mut db, &package_id, &host_id)?
         .as_bindings()
         .de()?
         .get(&internal_port)