+++ shared-libs/crates/start-core/src/service/effects/net/plugin.rs
@@ -94,18 +94,14 @@ pub async fn export_url(
         .ctx
         .db
         .mutate(|db| {
-            let host = host_for(
-                db,
-                hostname_info.package_id.as_ref(),
-                &hostname_info.host_id,
-            )?;
+            let host = host_for(db, &hostname_info.package_id, &hostname_info.host_id)?;
             host.as_bindings_mut()
                 .as_idx_mut(&hostname_info.internal_port)
                 .or_not_found(t!(
                     "net.plugin.binding-not-found",
                     binding = format!(
                         "{}:{}:{}",
-                        hostname_info.package_id.as_deref().unwrap_or("STARTOS"),
+                        hostname_info.package_id,
                         hostname_info.host_id,
                         hostname_info.internal_port
                     )